Test for 64bit build

+++ libnpk/external/xxtea/xxtea.h
@@ -3,5 +3,5 @@
 void xxtea_encode(int* v, int* k);
 void xxtea_decode(int* v, int* k);
 
-void xxtea_encode_buffer(char* in_buffer, unsigned int in_size, int* key, int cipherRemains);
-void xxtea_decode_buffer(char* in_buffer, unsigned int in_size, int* key, int cipherRemains);
+void xxtea_encode_buffer(char* in_buffer, off_t in_size, int* key, int cipherRemains);
+void xxtea_decode_buffer(char* in_buffer, off_t in_size, int* key, int cipherRemains);
